extract async_scope and async_manual_reset_event
